feat: spreadsheetification of lottery export (#4228)
* feat: spreadsheetification of lottery export

* fix: updates for tests

* fix: updates per pr comments

* fix: adding test for coverage

* fix: removes un ranked applications from export
