+++ api/package.json
@@ -67,6 +67,7 @@
     "compression": "^1.7.4",
     "cookie-parser": "~1.4.6",
     "dayjs": "~1.11.9",
+    "exceljs": "^4.4.0",
     "handlebars": "~4.7.8",
     "jsonwebtoken": "~9.0.1",
     "juice": "^9.1.0",
@@ -92,6 +93,7 @@
     "@nestjs/schematics": "^10.1.1",
     "@nestjs/testing": "^10.3.2",
     "@types/compression": "^1.7.5",
+    "@types/exceljs": "1.3.0",
     "@types/express": "~4.17.17",
     "@types/jest": "~29.5.3",
     "@types/multer": "^1.4.7",
