+++ sites/partners/src/lib/hooks.ts
@@ -522,12 +522,39 @@ export const useApplicationsExport = (listingId: string, includeDemographics: bo
 }
 
 export const useLotteryExport = (listingId: string) => {
-  const { applicationsService } = useContext(AuthContext)
+  const { lotteryService } = useContext(AuthContext)
+  const [exportLoading, setExportLoading] = useState(false)
+  const { addToast } = useContext(MessageContext)
 
-  return useCsvExport(
-    () => applicationsService.lotteryResults({ listingId, timeZone: dayjs.tz.guess() }),
-    `lottery-${listingId}-${createDateStringFromNow()}.csv`
-  )
+  const onExport = useCallback(async () => {
+    setExportLoading(true)
+    try {
+      const content = await lotteryService.lotteryResults(
+        { listingId },
+        { responseType: "arraybuffer" }
+      )
+      const blob = new Blob([new Uint8Array(content)], { type: "application/zip" })
+      const url = window.URL.createObjectURL(blob)
+      const link = document.createElement("a")
+      link.href = url
+      const now = new Date()
+      const dateString = dayjs(now).format("YYYY-MM-DD_HH-mm")
+      link.setAttribute("download", `${listingId}-${dateString}-test.zip`)
+      document.body.appendChild(link)
+      link.click()
+      link.parentNode.removeChild(link)
+      addToast(t("t.exportSuccess"), { variant: "success" })
+    } catch (err) {
+      console.log(err)
+      addToast(t("account.settings.alerts.genericError"), { variant: "alert" })
+    }
+    setExportLoading(false)
+  }, [])
+
+  return {
+    onExport,
+    exportLoading,
+  }
 }
 
 export const useUsersExport = () => {
@@ -539,7 +566,11 @@ export const useUsersExport = () => {
   )
 }
 
-const useCsvExport = (endpoint: () => Promise<string>, fileName: string) => {
+const useCsvExport = (
+  endpoint: () => Promise<string>,
+  fileName: string,
+  exportedAsSpreadsheet = false
+) => {
   const [csvExportLoading, setCsvExportLoading] = useState(false)
   const { addToast } = useContext(MessageContext)
 
@@ -548,7 +579,7 @@ const useCsvExport = (endpoint: () => Promise<string>, fileName: string) => {
 
     try {
       const content = await endpoint()
-      const blob = new Blob([content], { type: "text/csv" })
+      const blob = new Blob([content], { type: exportedAsSpreadsheet ? "text/xlsx" : "text/csv" })
       const fileLink = document.createElement("a")
       fileLink.setAttribute("download", fileName)
       fileLink.href = URL.createObjectURL(blob)
