+++ sites/partners/src/pages/listings/[id]/lottery.tsx
@@ -12,7 +12,6 @@ import { CardHeader, CardSection } from "@bloom-housing/ui-seeds/src/blocks/Card
 import { AuthContext } from "@bloom-housing/shared-helpers"
 import {
   Listing,
-  ListingUpdate,
   ListingEventsTypeEnum,
   ListingsStatusEnum,
   LotteryStatusEnum,
@@ -46,7 +45,7 @@ const Lottery = (props: { listing: Listing }) => {
   const [loading, setLoading] = useState(false)
 
   const { listingsService, lotteryService, profile } = useContext(AuthContext)
-  const { onExport, csvExportLoading } = useLotteryExport(listing?.id)
+  const { onExport, exportLoading } = useLotteryExport(listing?.id)
   const { data } = useFlaggedApplicationsMeta(listing?.id)
   const duplicatesExist = data?.totalPendingCount > 0
   let formattedExpiryDate: string
@@ -86,7 +85,7 @@ const Lottery = (props: { listing: Listing }) => {
         </div>
         <div>
           <Button
-            disabled={loading || csvExportLoading}
+            disabled={loading || exportLoading}
             onClick={() => {
               if (profile?.userRoles?.isAdmin) {
                 setExportModal(true)
@@ -121,7 +120,7 @@ const Lottery = (props: { listing: Listing }) => {
                 onClick={() => {
                   setRunModal(true)
                 }}
-                disabled={loading || csvExportLoading}
+                disabled={loading || exportLoading}
               >
                 {t("listings.lottery.runLottery")}
               </Button>
@@ -389,7 +388,7 @@ const Lottery = (props: { listing: Listing }) => {
                   }
                 }}
                 size="sm"
-                loadingMessage={loading || csvExportLoading ? t("t.loading") : undefined}
+                loadingMessage={loading || exportLoading ? t("t.loading") : undefined}
               >
                 {duplicatesExist
                   ? t("listings.lottery.runLotteryDuplicates")
@@ -591,7 +590,7 @@ const Lottery = (props: { listing: Listing }) => {
                   setExportModal(false)
                 }}
                 size="sm"
-                loadingMessage={loading || csvExportLoading ? t("t.loading") : undefined}
+                loadingMessage={loading || exportLoading ? t("t.loading") : undefined}
               >
                 {t("t.export")}
               </Button>
